Issue #4 - Move to stub/mock based testing.
The currently library is strongly coupled HA-Proxy, meaning that
in order to execute tests haproxy must be running locally. This
change adds an interface (HAProxy) that allows us to then stub out
the calls to RunCommand.

Note: If you're to go this route, README will need to be updated as
the Info and Stats methods are no longer pointer receivers, but instead
receiving the HAProxyClient as an argument. For example

    client := HAProxyClient{Addr: "tcp://blah:64738"}
    info, err := Info(client) {
        ...
    }
    ...